requirements.txt
flask==3.0.3
tensorflow==2.16.1
numpy==1.26.4
pillow==10.3.0
matplotlib==3.8.4
scikit-learn==1.4.2
werkzeug==3.0.2
h5py==3.11.0